lesson12-9

+++ app/Http/Controllers/TasksController.php
@@ -18,11 +18,19 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+    
+        if(\Auth::check()){
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at','desc')->get();
+            
+            $data=[
+                'user' => $user,
+                'tasks' => $tasks,
+                ];
+        }
+        return view('tasks.index',$data);
         
-        return view('tasks.index',[
-            'tasks' => $tasks,
-            ]);
     }
 
     /**
@@ -53,6 +61,7 @@ class TasksController extends Controller
             ]);
         
         $task = new Task;
+        $task->user_id = \Auth::user()->id;
         $task->content = $request->content;
         $task->status = $request->status;
         $task->save();
@@ -70,9 +79,14 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
         
+        if(\Auth::user()->id===$task->user_id){
         return view('tasks.show',[
             'task' => $task, 
             ]);
+        }
+        else{
+            return redirect('/');
+        }
        
     }
 
@@ -86,9 +100,14 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
         
+        if(\Auth::user()->id===$task->user_id){
         return view('tasks.edit',[
             'task' => $task,
             ]);
+        }
+        else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -107,11 +126,17 @@ class TasksController extends Controller
         
         
         $task = Task::find($id);
+        
+        if(\Auth::user()->id===$task->user_id){
         $task->content = $request->content;
         $task->status = $request->status;
         $task->save();
         
         return redirect('/');
+        }
+        else{
+        return redirect('/');
+        }
     }
 
     /**
@@ -123,8 +148,13 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
         
+        if(\Auth::user()->id===$task->user_id){
+        $task->delete();
         return redirect('/');
+        }
+        else{
+        return redirect('/'); 
+        }
     }
 }
